Handle newed variables and add stack trace (#39)
* Handle errors when %ut is newed

This makes the framework handle and display errors that happen in routines
that are called from a test but where the %ut variable is newed and thus
not available to the error trap code. In that case the error trap crashed
and did not display information about the original error.

This change updates the error trap to unwind the stack before handling the
error so that %ut is available and the original error can be displayed
properly to the user.

The BADERRORINNER is updated to test that this works as expected.

* Print stack trace on error

Updates the error trap to print a stack trace which shows where in the
code that the error occurred.

* Increment file version

---------

Co-authored-by: Jens Lidestrom <jens.lidestrom@vgregion.se>